Homework10/requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
pillow

conda install torch torchvision numpy matplotlib tqdm pillow